requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==39.0.0
defusedxml==0.7.1
Django==3.2.16
django-allauth==0.52.0
django-dump-load-utf8==0.0.4
django-environ==0.10.0
django-extensions==3.2.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
flake8==6.0.0
idna==3.4
isort==5.11.4
kombu==5.2.4
mccabe==0.7.0
oauthlib==3.2.2
Pillow==9.4.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7
redis==4.4.2
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
stripe==5.4.0
tzdata==2022.7
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
